==> 02-HISAT2/website/14_anova_kmeans.Rmd <==
---
title: "ANOVA and k-means clustering"
---

<!-- As HTML pages are built independently, large objects need to be
reimported from disk, and smaller ones quickly recomputed. -->

```{r setup, include=FALSE}
knitr::opts_chunk$set(cache = TRUE)
library(scater)
library(broom)
library(ggplot2)
library(RColorBrewer)
library(ComplexHeatmap)
library(reshape2)
sce.norm <- readRDS("rds/sce.norm.tSNE.rds")
```

# Preprocessing

Let us subset the `SCESet` to only endogenous features
(*i.e.*, exclude ERCC spike-in features):

```{r sce.endo}
sce.endo <- sce.norm[!isSpike(sce.norm)]
```

Let us first define:

* a set of colours used to indicate *cluster membership*,
  *Infection* and *Status* phenotypes consistently in the upcoming sections:

```{r col.sets}
col.set3 <- brewer.pal(12, "Set3")
col.cluster <- col.set3[c(1,12,3,4,9)]; names(col.cluster) <- c(1:3,0,"NA")
col.status<-col.set3[c(3,5:6)];names(col.status)<-levels(sce.endo$Status)
col.infection<-col.set3[c(3,8,10)];names(col.infection)<-levels(sce.endo$Infection)
col.time<-col.set3[c(2,7:8)];names(col.time)<-levels(sce.endo$Time)
```

# Data subsets {.tabset}

## Mock - uninfected

```{r sce.mock}
sce.mock <- sce.endo[,sce.endo$Treatment == "Mock_uninfected"]
dim(sce.mock)
```

## D23580 infected

```{r sce.d23inf}
sce.d23inf <- sce.endo[,sce.endo$Treatment == "D23580_infected"]
dim(sce.d23inf)
```

## D23580 exposed

```{r sce.d23exp}
sce.d23exp <- sce.endo[,sce.endo$Treatment == "D23580_exposed"]
dim(sce.d23exp)
```

## LT2 infected

```{r sce.lt2inf}
sce.lt2inf <- sce.endo[,sce.endo$Treatment == "LT2_infected"]
dim(sce.lt2inf)
```

## LT2 exposed

```{r sce.lt2exp}
sce.lt2exp <- sce.endo[,sce.endo$Treatment == "LT2_exposed"]
dim(sce.lt2exp)
```

# One-way ANOVA {.tabset}

## D23580 infected

```{r aov.d23inf}
aov.d23inf <- do.call("rbind", apply(
  norm_exprs(sce.d23inf), 1, function(ne){
    tidy(anova(lm(ne ~ sce.d23inf$Time)))[1,2:6]
  }
))
aov.d23inf$gene_name <- fData(sce.d23inf)[,"gene_name"]
aov.d23inf$BH <- p.adjust(aov.d23inf$p.value, "BH")
sce.d23inf.aov <- sce.d23inf[aov.d23inf$BH < 0.01 & !is.na(aov.d23inf$BH),]
dim(sce.d23inf.aov)
```

## D23580 exposed

```{r aov.d23exp}
aov.d23exp <- do.call("rbind", apply(
  norm_exprs(sce.d23exp), 1, function(ne){
    tidy(anova(lm(ne ~ sce.d23exp$Time)))[1,2:6]
  }
))
aov.d23exp$gene_name <- fData(sce.d23exp)[,"gene_name"]
aov.d23exp$BH <- p.adjust(aov.d23exp$p.value, "BH")
sce.d23exp.aov <- sce.d23exp[aov.d23exp$BH < 0.01 & !is.na(aov.d23exp$BH),]
dim(sce.d23exp.aov)
```

## LT2 infected

```{r aov.lt2inf}
aov.lt2inf <- do.call("rbind", apply(
  norm_exprs(sce.lt2inf), 1, function(ne){
    tidy(anova(lm(ne ~ sce.lt2inf$Time)))[1,2:6]
  }
))
aov.lt2inf$gene_name <- fData(sce.lt2inf)[,"gene_name"]
aov.lt2inf$BH <- p.adjust(aov.lt2inf$p.value, "BH")
sce.lt2inf.aov <- sce.lt2inf[aov.lt2inf$BH < 0.01 & !is.na(aov.lt2inf$BH),]
dim(sce.lt2inf.aov)
```

## LT2 exposed

```{r aov.lt2exp}
aov.lt2exp <- do.call("rbind", apply(
  norm_exprs(sce.lt2exp), 1, function(ne){
    tidy(anova(lm(ne ~ sce.lt2exp$Time)))[1,2:6]
  }
))
aov.lt2exp$gene_name <- fData(sce.lt2exp)[,"gene_name"]
aov.lt2exp$BH <- p.adjust(aov.lt2exp$p.value, "BH")
sce.lt2exp.aov <- sce.lt2exp[aov.lt2exp$BH < 0.01 & !is.na(aov.lt2exp$BH),]
dim(sce.lt2exp.aov)
```

# *P*-value

```{r ggplot_p.value}
ggplot(data.frame(rbind(
  cbind(aov.d23inf, Treatment = "D23580 infected"),
  cbind(aov.d23exp, Treatment = "D23580 exposed"),
  cbind(aov.lt2inf, Treatment = "LT2 infected"),
  cbind(aov.lt2exp, Treatment = "LT2 exposed")
))) +
  geom_density(aes(-log10(p.value), colour = Treatment)) +
  theme_minimal()
```

# k-means clustering {.tabset}

## D23580 infected

```{r km.d23inf}
km.d23inf <- kmeans(norm_exprs(sce.d23inf.aov), 9)
table(km.d23inf$cluster)
```

## D23580 exposed

```{r km.d23exp}
km.d23exp <- kmeans(norm_exprs(sce.d23exp.aov), 9)
table(km.d23exp$cluster)
```

## LT2 infected

```{r km.lt2inf}
km.lt2inf <- kmeans(norm_exprs(sce.lt2inf.aov), 9)
table(km.lt2inf$cluster)
```

## LT2 exposed

```{r km.lt2exp}
km.lt2exp <- kmeans(norm_exprs(sce.lt2exp.aov), 9)
table(km.lt2exp$cluster)
```

# Clusters {.tabset}

## D23580 infected

```{r meanTime.d23inf, echo=FALSE}
meanTime.d23inf <- melt(apply(
  norm_exprs(sce.d23inf.aov), 1, function(ne){
    tapply(ne, sce.d23inf.aov$Time, mean)
  }
))
colnames(meanTime.d23inf) <- c("Time", "Gene", "norm_exprs")
meanTime.d23inf$Cluster <- sprintf(
  "cluster %i\n%i genes",
  km.d23inf$cluster[meanTime.d23inf$Gene],
  table(km.d23inf$cluster)[km.d23inf$cluster[meanTime.d23inf$Gene]]
)
clusterOrder <- with(
  dcast(
    aggregate(norm_exprs ~ Cluster + Time, meanTime.d23inf, "mean"),
    Cluster ~ Time, value.var = "norm_exprs"
  ),
  Cluster[order(`2h`,`4h`,`6h`)])
meanTime.d23inf$Cluster <- factor(meanTime.d23inf$Cluster, clusterOrder)
ggplot(meanTime.d23inf) +
  geom_line(aes(Time, norm_exprs, group = Gene), alpha = 0.1) +
  facet_wrap(~ Cluster) +
  labs(
    title = "D23580 infected",
    subtitle = sprintf("%i genes", length(km.d23inf$cluster))
  ) +
  scale_y_continuous(limits = range(norm_exprs(sce.endo))) +
  theme_minimal()
ggsave("14_out/meanTime.D23580_infected.pdf", width = 6, height = 5)
```

```{r d23.c1, echo=FALSE, eval=FALSE}
d23.c1 <- names(which(km.d23inf$cluster == 1))
View(subset(fData(sce.d23inf), gene_id %in% d23.c1, "gene_name"))
intersect(
  names(which(km.d23inf$cluster == 1)),
  names(which(km.lt2inf$cluster == 9))
)
```

## D23580 exposed

```{r meanTime.d23exp, echo=FALSE}
meanTime.d23exp <- melt(apply(
  norm_exprs(sce.d23exp.aov), 1, function(ne){
    tapply(ne, sce.d23exp.aov$Time, mean)
  }
))
colnames(meanTime.d23exp) <- c("Time", "Gene", "norm_exprs")
meanTime.d23exp$Cluster <- sprintf(
  "cluster %i\n%i genes",
  km.d23exp$cluster[meanTime.d23exp$Gene],
  table(km.d23exp$cluster)[km.d23exp$cluster[meanTime.d23exp$Gene]]
)
clusterOrder <- with(
  dcast(
    aggregate(norm_exprs ~ Cluster + Time, meanTime.d23exp, "mean"),
    Cluster ~ Time, value.var = "norm_exprs"
  ),
  Cluster[order(`2h`,`4h`,`6h`)])
meanTime.d23exp$Cluster <- factor(meanTime.d23exp$Cluster, clusterOrder)
ggplot(meanTime.d23exp) +
  geom_line(aes(Time, norm_exprs, group = Gene), alpha = 0.1) +
  facet_wrap(~ Cluster) +
  labs(
    title = "D23580 exposed",
    subtitle = sprintf("%i genes", length(km.d23exp$cluster))
  ) +
  scale_y_continuous(limits = range(norm_exprs(sce.endo))) +
  theme_minimal()
ggsave("14_out/meanTime.D23580_exposed.pdf", width = 6, height = 5)
```


## LT2 infected

```{r meanTime.lt2inf, echo=FALSE}
meanTime.lt2inf <- melt(apply(
  norm_exprs(sce.lt2inf.aov), 1, function(ne){
    tapply(ne, sce.lt2inf.aov$Time, mean)
  }
))
colnames(meanTime.lt2inf) <- c("Time", "Gene", "norm_exprs")
meanTime.lt2inf$Cluster <- sprintf(
  "cluster %i\n%i genes",
  km.lt2inf$cluster[meanTime.lt2inf$Gene],
  table(km.lt2inf$cluster)[km.lt2inf$cluster[meanTime.lt2inf$Gene]]
)
clusterOrder <- with(
  dcast(
    aggregate(norm_exprs ~ Cluster + Time, meanTime.lt2inf, "mean"),
    Cluster ~ Time, value.var = "norm_exprs"
  ),
  Cluster[order(`2h`,`4h`,`6h`)])
meanTime.lt2inf$Cluster <- factor(meanTime.lt2inf$Cluster, clusterOrder)
ggplot(meanTime.lt2inf) +
  geom_line(aes(Time, norm_exprs, group = Gene), alpha = 0.1) +
  facet_wrap(~ Cluster) +
  labs(
    title = "LT2 infected",
    subtitle = sprintf("%i genes", length(km.lt2inf$cluster))
  ) +
  scale_y_continuous(limits = range(norm_exprs(sce.endo))) +
  theme_minimal()
ggsave("14_out/meanTime.LT2_infected.pdf", width = 6, height = 5)
```

## LT2 exposed

```{r meanTime.lt2exp, echo=FALSE}
meanTime.lt2exp <- melt(apply(
  norm_exprs(sce.lt2exp.aov), 1, function(ne){
    tapply(ne, sce.lt2exp.aov$Time, mean)
  }
))
colnames(meanTime.lt2exp) <- c("Time", "Gene", "norm_exprs")
meanTime.lt2exp$Cluster <- sprintf(
  "cluster %i\n%i genes",
  km.lt2exp$cluster[meanTime.lt2exp$Gene],
  table(km.lt2exp$cluster)[km.lt2exp$cluster[meanTime.lt2exp$Gene]]
)
clusterOrder <- with(
  dcast(
    aggregate(norm_exprs ~ Cluster + Time, meanTime.lt2exp, "mean"),
    Cluster ~ Time, value.var = "norm_exprs"
  ),
  Cluster[order(`2h`,`4h`,`6h`)])
meanTime.lt2exp$Cluster <- factor(meanTime.lt2exp$Cluster, clusterOrder)
ggplot(meanTime.lt2exp) +
  geom_line(aes(Time, norm_exprs, group = Gene), alpha = 0.1) +
  facet_wrap(~ Cluster) +
  labs(
    title = "LT2 exposed",
    subtitle = sprintf("%i genes", length(km.lt2exp$cluster))
  ) +
  scale_y_continuous(limits = range(norm_exprs(sce.endo))) +
  theme_minimal()
ggsave("14_out/meanTime.LT2_exposed.pdf", width = 6, height = 5)
```

# Legacy code: to do

Heat map of ANOVA-significant genes:

* consider indicating cluster membership

```{r Heatmap_d23, eval=FALSE}
cellOrder <- order(sce.d23.aov$Time)
ha <- HeatmapAnnotation(
  df = pData(sce.d23.aov)[cellOrder, "Time", drop = FALSE],
  col = list(
    Time = col.time
  )
)
hm <- Heatmap(
  norm_exprs(sce.d23.aov)[,cellOrder],
  name = "norm_exprs", column_title = "norm_exprs", cluster_columns = FALSE,
  top_annotation = ha, show_row_names = FALSE, show_column_names = FALSE
)
draw(hm)
```
