==> 02-HISAT2/website/99_Bonus.Rmd <==

---
title: "Bonus"
bibliography:
  bibtex.bib
---

Enjoy a few by-products of this analysis!

# Songs

## Single-cell

*Guided by the flow*

*Into separate wells*

... Down the tube they go /  Lysed by the buffer ...

-- --

*Single cells, single cells,*

*Single all the way*

*Oh, what fun it is to yield*

*So little RNA, hey!*

-- --


## Wake me up when sequencing ends

*Samples have come and passed*

*The quality can never last*

*Wake me up when sequencing ends*


*Validation's come to pass*

*Seven years have gone so fast*

*Wake me up when sequencing ends*


*Here come the funds again*

*Falling from the stars*

*Drenched in review again*

*Telling what we have*


*Now my article's out*

*Never forget what I claimed*

*Wake me up when sequencing ends*


*Samples have come and passed*

*The quality can never last*

*Wake me up when sequencing ends*


...(Ring out the bells again)
...(Like we did when spring began)
*Wake me up when sequencing ends*


*Here come the funds again*

*Falling from the stars*

*Drenched in review again*

*Telling what we have*


*Now my article's out*

*Never forget what I claimed*

*Wake me up when sequencing ends*


*Samples have come and passed*

*The quality can never last*

*Wake me up when sequencing ends*


*Validation's come to pass*

*Seven years have gone so fast*

*Wake me up when sequencing ends* (x3)
